+++ terraform/modules/essbase/outputs.tf
@@ -1,16 +1,12 @@
 ## Copyright (c) 2019, 2020, Oracle and/or its affiliates.
 ## Licensed under the Universal Permissive License v1.0 as shown at http://oss.oracle.com/licenses/upl.
 
-output "node_name" {
-  value = oci_core_instance.essbase.display_name
-}
-
 output "node_id" {
   value = oci_core_instance.essbase.id
 }
 
-output "node_domain_name" {
-  value = local.node_domain_name
+output "node_display_name" {
+  value = oci_core_instance.essbase.display_name
 }
 
 output "node_public_ip" {