+++ terraform/provider.tf
@@ -29,7 +29,7 @@ variable "private_key_path" {
 }
 
 provider "oci" {
-  version = "~> 3.53.0"
+  version = "~> 3.68.0"
   tenancy_ocid = var.tenancy_ocid
   region       = var.region
   user_ocid        = var.user_ocid