+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Support for private load balancer configuration for internal access.
+
+### Changed
+- Bastion host is now an optional component.
+- Essbase stack configuration is started in the background, which eliminates the need for the bastion host in private environments.
+- Reorganized the layout of the guided UI.
+
 ## [19.3.0.0.3]
 
 ### Changed